fix: move expected supplier to test fixtures
