+++ test/fixtures/contentful-supplier.js
